rename images/homes to images/home
